+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 3.0.0
+
+- Support for processing non-implied resource file
+
+   > - non-implied resource file:  the resource file which is outside of `lib/` directory, for example:
+   >    - `~/path/to/flutter_r_demo/assets/images/test.png` 
+   >    - `~/path/to/flutter_r_demo/assets/images/3.0x/test.png`
+   > - implied resource file:  the resource file which is inside of  `lib/` directory, for example:
+   >    - `~/path/to/flutter_r_demo/lib/assets/images/hot_foot_N.png` 
+   >    - `~/path/to/flutter_r_demo/lib/assets/images/3.0x/hot_foot_N.png`
+
+- New recommended flutter resource structure
+
 ## 2.0.0
 
 - New asset generation algorithm to support all kinds of standard or nonstandard image/text resource structure